style: 달력 칸 형식 조절
