+++ src/styles/CalendarStyles.jsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const CalendarTile = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 18px;
+  font-size: 12px;
+  color: #555;
+`;
